+++ tests/test_async_trait.rs
@@ -9,7 +9,7 @@ async fn foo<A: Bar>(a: &A) -> u32 {
     a.bar().await
 }
 
-#[entrait(Bar, async_trait = true, unimock = true)]
+#[entrait(Bar, async_trait, unimock)]
 async fn bar(state: &State) -> u32 {
     state.0
 }