+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- no_deps support. Add this attribute to not interpret the first parameter as a deps parameter.
+- default values for config attributes (may skip '= value')
 
 ## [0.3.0] - 2022-06-03
 ### Changed